[FLINK-22053][core] Allow NumberSequenceSource to have less splits than parallelism.
